pyproject.toml
[project]
name = "md-to-docx"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "docling>=2.69.0",
    "mammoth>=1.11.0",
    "markdown-it-py>=4.0.0",
    "pypandoc>=1.16.2",
    "python-docx>=1.2.0",
    "pydantic>=2.0.0",
    "httpx>=0.25.0",
    "lxml>=5.0.0",
]
